Fix bytea_decode_in_place.
So apparently the sign of a char is more important than I thought it was for these calculations. It is now fully working as far as I can tell. I had changed the test, but JS doesn't know what to do with char values > 127 so I put it back for now.